imt4641 computational forensics
Abstract:

Mobile forensics is a relatively new thing, but with the increasing
amount of mobile phones,
tablets and applications created for these, a forensic investigator
needs more knowledge about
these devices. There are several hundred thousand applications out
there, each created for a
dierent OS with a dierent storage system (e.g. XML, CSV, SQLite, Text
les, etc.).
How should an investigator proceed to get the needed evidence in time?
What processes can be
used to speed up the investigation? What should an investigator look
for? My focus on this
problem will be to attempt to develop an automation tool for the Android
OS, by use of machine
learning. A prerequisite for using the tool would be that an image of
the device(s) in question is
already made, and the SQLite databases extracted from the phone for
analysis. The end goal for
the tool to be developed is the possibility to analyze the le storage
(SQLite) and retrieve the data
the investigator needs. The main focus will be on logs and information
retrieval, like image data,
GPS/localization logs, e-mail information or other kinds of logs/data
found in the database(s).